requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.5
PyYAML>=6.0
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0 
doclayout-yolo  
transformers>=4.30.0  
huggingface-hub>=0.16.0
pytesseract>=0.3.10  
easyocr>=1.7.0  
python-doctr[torch]>=0.8.0  
openai>=1.0.0  
chromadb>=0.4.0
sentence-transformers>=2.2.0
pdf2image>=1.16.0
pypdf2>=3.0.0
tqdm>=4.64.0
python-dotenv>=0.19.0  
gradio>=4.0.0 
